requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
web3>=7.0.0
chaoschain-sdk>=0.4.0
x402>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.50.0
apscheduler>=3.10.0
